+++ src/view/animation.js
